Extract OneFileProcessor to files_worker module
Refactor by moving OneFileProcessor class from worker.py to dedicated files_worker.py module. This improves separation of concerns and makes the codebase more modular. Update imports in worker.py and add test file stub for files_worker.
