+++ src/main_app/jobs_workers/public_jobs_workers/copy_svg_langs/files_worker.py
@@ -1,0 +1,361 @@
+"""
+Worker module for copy_svg_langs.
+"""
+
+from __future__ import annotations
+
+import logging
+from pathlib import Path
+
+from ....api_services.files_service import FilesService
+from ....shared.copysvg_wrapper import (
+    InjectResult,
+    MatchFixNestedTags,
+    inject_step_one_file,
+)
+from ....shared.copysvg_wrapper.mapping import ExtractorData
+from ....shared.fix_nested import (
+    DetectionResult,
+    VerificationResult,
+)
+from .objects import (
+    FilesProcessedItem,
+    StepResult,
+    SvgLangsConfig,
+)
+
+logger = logging.getLogger(__name__)
+
+
+class OneFileProcessor:
+
+    def __init__(self, config: SvgLangsConfig, files_service: FilesService) -> None:
+        self.config = config
+        self.files_service = files_service
+        self.mapping: ExtractorData | None = None
+        self.upload_done = 0
+        self.nested_processer: MatchFixNestedTags
+
+    def _set_translations(self, mapping: ExtractorData) -> None:
+        self.mapping = mapping
+
+    def _update_translations(self, mapping: ExtractorData) -> None:
+        if self.mapping is None:
+            return self._set_translations(mapping)
+
+        self.mapping.merge(mapping, merge_keys=["new", "title_new"])
+
+    def _process_one_item(self, title: str, title_info: FilesProcessedItem, main_title: str) -> bool:
+        # ----------------------------------------------
+        # File step 1: download
+
+        file_path_str: str | None = self.get_download_path(title, title_info)
+
+        if not file_path_str:
+            return False
+
+        title_info.file_path = file_path_str
+
+        file_path = Path(file_path_str)
+
+        # ----------------------------------------------
+        # File step 2: fix nested tags
+        nested_step = title_info.steps.nested
+
+        verify_fixed, no_nested_tags = self.handle_nested_tag_repair_step(nested_step, file_path)
+
+        if not no_nested_tags:
+            # no nested tags fixed, break the file process
+            title_info.status = "failed"
+
+            # We can't inject file that has nested tags
+            title_info.steps.inject.msg = "skipped"
+            title_info.steps.upload.msg = "skipped"
+            return False
+
+        # ----------------------------------------------
+        # At this point, no nested tags remaining in the file
+        # File step 3: log translations
+        # File step 4: inject translations
+
+        new_path: Path | None = self.inject_step_file(title_info, file_path)
+        inject_result = title_info.steps.inject
+
+        # ----------------------------------------------
+        # File step 5: upload
+        if inject_result.result is True:
+            # inject success
+            translation_details = title_info.steps.translations.details or {}
+            summary = self._create_language_summary(main_title, translation_details)
+            return self._upload_step(title_info, summary, new_path)
+
+        # ----------------------------------------------
+        if verify_fixed > 0:
+            # Here we need to upload the orignal file because we fix nested tags.
+            summary = f"{verify_fixed} nested tags fixed"
+            return self._upload_step(title_info, summary, file_path)
+
+        # No nested tags were fixed
+        # inject_result.result is one of (None/False)
+        return False
+
+    def handle_nested_tag_repair_step(self, nested_step: StepResult, file_path: Path) -> tuple[int, bool]:
+        nested_processer = MatchFixNestedTags(
+            source_file=file_path,
+            new_path=file_path,
+        )
+
+        detect_before: DetectionResult = nested_processer.detect_nested_tags()
+        if detect_before.count == 0:
+            nested_step._update(msg="No nested tags found")
+            # no nested tags, process to inject translations step
+            return 0, True
+
+        # Try to fix nested tags
+        if not nested_processer.fix_file():
+            nested_step._update(
+                result=False,
+                msg="Failed to fix nested tags",
+                details=detect_before.to_dict(),
+            )
+            # no nested tags fixed, break the file process
+            return 0, False
+
+        verify: VerificationResult = nested_processer.verify_after_fix()
+
+        if verify.fixed == 0:
+            nested_step._update(
+                result=False,
+                msg="No nested tags were fixed",
+                details=verify.to_dict(),
+            )
+            # no nested tags fixed, break the file process
+            return 0, False
+
+        verify_fixed = verify.fixed
+
+        nested_step._update(
+            result=True,
+            msg=f"Fixed {verify.fixed} nested tag(s)",
+            details=verify.to_dict(),
+        )
+
+        # no nested tags remaining in the file, process to inject translations step
+        return verify_fixed, True
+
+    def inject_step_file(self, title_info: FilesProcessedItem, file_path: Path | str) -> Path | None:
+        if not file_path:
+            title_info.steps.inject._update(result=False, msg="No file path found")
+            return None
+
+        file_path = Path(file_path)
+        output_file = self.config.output_dir / "translated" / file_path.name
+
+        inject_result: InjectResult = inject_step_one_file(
+            file_path=file_path,
+            translations=self.mapping,
+            output_file=output_file,
+            overwrite_translations=self.config.overwrite_translations,
+        )
+
+        if inject_result.result is None:
+            if inject_result.msg == "No changes":
+                title_info.status = "skipped"
+
+            title_info.steps.inject._update(
+                result=None,
+                msg=inject_result.msg,
+            )
+            return None
+
+        if inject_result.result is False:
+            title_info.steps.inject._update(
+                result=False,
+                msg=inject_result.msg,
+            )
+            return None
+
+        if inject_result.result is not True:
+            title_info.steps.inject._update(
+                result=False,
+                msg="Unknown error",
+            )
+            return None
+
+        # inject_result.result is True definitely
+        title_info.steps.translations._update(
+            result=True,
+            details={
+                "new_list": inject_result.languages_after,
+                "new": inject_result.new_languages_count or 0,
+                "updated": inject_result.updated_translations or 0,
+                "inserted": inject_result.inserted_translations or 0,
+            },
+        )
+
+        title_info.steps.inject._update(
+            result=True,
+            msg=inject_result.msg,
+            # details={
+            #     "new_languages_count": inject_result.new_languages_count,
+            #     "updated_translations": inject_result.updated_translations,
+            #     "output_file": output_file,
+            # },
+        )
+
+        return output_file
+
+    def _upload_step(
+        self,
+        title_info: FilesProcessedItem,
+        summary: str,
+        new_path: Path | None,
+    ) -> bool:
+        # Check if settings upload_files option is disabled
+        if self.config.upload_files is False:
+            title_info.steps.upload._update(
+                result=None,
+                msg="skipped",
+                details={"error": "Upload disabled from settings"},
+            )
+            title_info.status = "skipped"
+            return False
+
+        # Check if form upload input is enabled
+        if not bool(self.config.upload):
+            title_info.steps.upload._update(
+                result=None,
+                msg="skipped",
+                details={"error": "Upload disabled"},
+            )
+            title_info.status = "skipped"
+            return False
+
+        if self.config.upload_limit > 0 and self.upload_done >= self.config.upload_limit:
+            title_info.steps.upload._update(
+                result=None,
+                msg="skipped",
+                details={"error": "Upload limit reached"},
+            )
+            title_info.status = "skipped"
+            return False
+
+        # Start uploading
+        upload = self.files_service.upload_fixed_svg(
+            title_info.title,
+            new_path,
+            summary=summary,
+        )
+        upload_success = upload.get("ok")
+        upload_error = upload.get("error") or ""
+        upload_msg = upload.get("msg") or ""
+        error_details = upload.get("error_details", "")
+
+        if upload_success is True:
+            title_info.steps.upload._update(
+                result=True,
+                msg="File Successfully uploaded.",
+                # details=upload.get("result", ""),
+            )
+
+            self.upload_done += 1
+            title_info.status = "success"
+            # return True, all steps passed and upload is success
+            return True
+
+        error_and_details = {
+            "error": upload_error,
+            "error_details": error_details,
+        }
+
+        is_no_changes = upload_error in {"skipped", "fileexists-no-change"}
+        if upload_success is None and is_no_changes:
+            title_info.steps.upload._update(
+                result=None,
+                msg=upload_msg,
+                details=error_and_details,
+            )
+            title_info.status = "skipped"
+            return False
+
+        title_info.error = upload_error
+        title_info.steps.upload._update(
+            result=False,
+            msg="Upload failed.",
+            details=error_and_details,
+        )
+        title_info.status = "failed"
+        return False
+
+    def _create_language_summary(self, main_title: str, translation_details: dict[str, int]) -> str:
+        new_count = translation_details.get("new", 0)
+        updated = translation_details.get("updated", 0)
+        inserted = translation_details.get("inserted", 0)
+
+        file_name = main_title.removeprefix("File:")
+
+        summary_list = []
+        if new_count > 0:
+            summary_list.append(f"{new_count} languages injected")
+
+        if updated > 0:
+            summary_list.append(f"{updated} translations Updated")
+
+        if inserted > 0:
+            summary_list.append(f"{inserted} translations inserted")
+
+        if not summary_list:
+            summary_list.append("Adding translations")
+
+        summary = ", ".join(summary_list)
+
+        summary += f" from [[File:{file_name}]]"
+
+        return summary
+
+    def get_download_path(self, title: str, title_info: FilesProcessedItem):
+        down_step = title_info.steps.download
+        try:
+            file_data = self.files_service.download_one_file(
+                title=title,
+                out_dir=self.config.output_dir_files,
+                overwrite_download=self.config.overwrite_download,
+            )
+        except Exception as e:
+            logger.exception("Error downloading SVG file")
+            down_step._update(result=False, msg="Error downloading", details={"error": str(e)})
+            title_info.status = "failed"
+            return None
+
+        if file_data.get("result") != "success":
+            download_result = {
+                "ok": False,
+                "path": None,
+                "error": "download_failed",
+                "details": file_data,
+            }
+            down_step._update(result=False, msg="Failed to download file", details=download_result)
+            title_info.status = "failed"
+            return None
+
+        file_path: str | None = file_data.get("path")
+
+        download_result = {
+            "ok": True,
+            "path": file_path,
+            "error": None,
+            "details": {},
+        }
+
+        if file_path:
+            down_step._update(result=True, msg="Downloaded successfully", details=download_result)
+            return file_path
+
+        down_step._update(result=False, msg="Failed to get file path", details=download_result)
+        title_info.status = "failed"
+        return None
+
+
+__all__ = [
+    "OneFileProcessor",
+]
