+++ tests/unit/jobs_workers/public_jobs_workers/copy_svg_langs/test_files_worker.py
@@ -1,0 +1,16 @@
+# ruff: noqa: F401
+"""
+Unit tests for files_worker module.
+
+classes to test: OneFileProcessor
+
+TODO: write tests
+"""
+
+from __future__ import annotations
+
+import pytest
+
+from src.main_app.jobs_workers.public_jobs_workers.copy_svg_langs.files_worker import (
+    OneFileProcessor,
+)
